Saving child > parent has one
The parent child relationship was being set but not the child > parent relationship. 

Example: I could query $this->Child->ID on the parent object but not $this->Partent->ID on the child object.

I am sure there is a more 'ORM' way to do this but this worked for my situation.